+++ geotrek/altimetry/tests/test_views.py
@@ -1,6 +1,10 @@
-from django.test import TestCase
+from django.contrib.gis.geos import LineString
+from django.db import connection
+from django.test import TestCase, override_settings
 
+from geotrek import settings
 from geotrek.authent.tests.factories import UserFactory
+from geotrek.core.tests.factories import PathFactory
 from geotrek.trekking.tests.factories import TrekFactory
 
 
@@ -32,3 +36,54 @@ class ProfileViewsTest(TestCase):
         trek = TrekFactory.create(name='Trek', published=True)
         response = self.client.get('/media/profiles/trek-%s.png' % trek.pk)
         self.assertEqual(response.status_code, 200)
+
+
+class ProfileCacheTests(TestCase):
+    """ Test profile svg is cached
+    """
+
+    TMP_CACHES = {
+        'default': {
+            'BACKEND': 'django.core.cache.backends.filebased.FileBasedCache',
+            'LOCATION': settings.CACHE_ROOT
+        }
+    }
+
+    @classmethod
+    def setUpTestData(cls):
+        # Create a simple fake DEM
+        with connection.cursor() as cur:
+            cur.execute('INSERT INTO altimetry_dem (rast) VALUES (ST_MakeEmptyRaster(100, 125, 0, 125, 25, -25, 0, 0, %s))', [settings.SRID])
+            cur.execute('UPDATE altimetry_dem SET rast = ST_AddBand(rast, \'16BSI\')')
+            demvalues = [[0, 0, 3, 5], [2, 2, 10, 15], [5, 15, 20, 25], [20, 25, 30, 35], [30, 35, 40, 45]]
+            for y in range(0, 5):
+                for x in range(0, 4):
+                    cur.execute('UPDATE altimetry_dem SET rast = ST_SetValue(rast, %s, %s, %s::float)', [x + 1, y + 1, demvalues[y][x]])
+        cls.path = PathFactory.create(geom=LineString((1, 101), (81, 101), (81, 99)))
+        cls.trek = TrekFactory.create(paths=[cls.path])
+
+    def test_cache_is_used_when_getting_trek_profile(self):
+        # There are 6 queries to get trek profile
+        with self.assertNumQueries(6):
+            response = self.client.get(f"/api/fr/treks/{self.trek.pk}/profile.json")
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response['Content-Type'], 'application/json')
+        # When cache is used there are only 2 queries to get trek profile
+        with self.assertNumQueries(2):
+            response = self.client.get(f"/api/fr/treks/{self.trek.pk}/profile.json")
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response['Content-Type'], 'application/json')
+
+    # Override default cache with fat cache since we can't use memcached in tests
+    @override_settings(CACHES=TMP_CACHES)
+    def test_cache_is_used_when_getting_trek_profile_svg(self):
+        # There are 6 queries to get trek profile svg
+        with self.assertNumQueries(6):
+            response = self.client.get(f"/api/fr/treks/{self.trek.pk}/profile.svg")
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response['Content-Type'], 'image/svg+xml')
+        # When cache is used there are only 7 queries to get trek profile
+        with self.assertNumQueries(5):
+            response = self.client.get(f"/api/fr/treks/{self.trek.pk}/profile.svg")
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response['Content-Type'], 'image/svg+xml')